+++ CDP4Reporting.Tests/DataCollection/DataCollectorCategoryColumnTestFixture.cs
@@ -23,7 +23,7 @@
 // </copyright>
 // --------------------------------------------------------------------------------------------------------------------
 
-namespace CDP4Composition.Tests.DataCollector
+namespace CDP4Reporting.Tests.DataCollection
 {
     using System;
     using System.Collections.Concurrent;
@@ -35,7 +35,7 @@ namespace CDP4Composition.Tests.DataCollector
     using CDP4Common.SiteDirectoryData;
     using CDP4Common.Types;
 
-    using CDP4Composition.DataCollector;
+    using CDP4Reporting.DataCollection;
 
     using NUnit.Framework;
