+++ CDP4Reporting/CDP4Reporting.csproj
@@ -12,6 +12,7 @@
     <Copyright>Copyright © RHEA System S.A.</Copyright>
     <AssemblyVersion>0.0.1.0</AssemblyVersion>
     <FileVersion>0.0.1.0</FileVersion>
+    <LangVersion>latest</LangVersion>
   </PropertyGroup>
   <PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Debug|AnyCPU' ">
     <DebugType>full</DebugType>
